+++ app/llm/langchain/__init__.py
@@ -1,3 +1,3 @@
 from llm.langchain.iris_langchain_completion_model import IrisLangchainCompletionModel
 from llm.langchain.iris_langchain_chat_model import IrisLangchainChatModel
-from llm.langchain.iris_langchain_embedding import IrisLangchainEmbeddingModel
+from llm.langchain.iris_langchain_embedding_model import IrisLangchainEmbeddingModel
